Update Playground.stories.tsx and Sandpack.stories.tsx
